more protocols
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@33662 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: 4839497555c23f5ed2772e462430281280dd6007

+++ src/router/opendpi/opendpi/include/ndpi_api.h
@@ -58,7 +58,7 @@ extern "C" {
   /**
    * This function returns the nDPI protocol id for IP-based protocol detection
    */
-	static u_int16_t ndpi_network_ptree_match(struct ndpi_detection_module_struct *ndpi_struct, struct in_addr *pin);
+	static u_int32_t ndpi_network_ptree_match(struct ndpi_detection_module_struct *ndpi_struct, struct in_addr *pin);
 
   /**
    * Same as ndpi_network_ptree_match
